Move ancient functions out of fetch into fetch_old
