+++ Ska/engarchive/fetch.py
@@ -1260,103 +1260,6 @@ class Msidset(MSIDset):
                                       filter_bad=filter_bad, stat=stat)
 
 
-def fetch_records(start, stop, msids, dt=32.8):
-    """
-    Fetch data from the telemetry archive as a recarray at a uniform time
-    spacing.
-
-    Only records where all columns have good quality get included.  This
-    routine can be substantially slower than fetch_arrays because of the
-    interpolation onto a common time sequence.
-
-    :param start: start date of telemetry (Chandra.Time compatible)
-    :param stop: stop date of telemetry
-    :param msids: list of MSIDs (case-insensitive)
-    :param dt: sampling time (sec)
-
-    :returns: numpy recarray with columns for time (CXC seconds) and ``msids``
-    """
-    import Ska.Numpy
-
-    with _cache_ft():
-        tstart, tstop, times, values, quals = _fetch(start, stop, msids)
-    dt_times = np.arange(tstart, tstop, dt)
-
-    len_times = len(dt_times)
-    bad = np.zeros(len_times, dtype=bool)
-    for msid in msids:
-        MSID = msid.upper()
-        index = Ska.Numpy.interpolate(np.arange(len(times[MSID])),
-                                      times[MSID], dt_times, 'nearest')
-        # Include also some interpolation validation check, something like (but
-        # not exactly) (abs(dt_times - times[MSID][index]) > 1.05 * dt)
-        bad |= quals[MSID][index]
-        values[msid] = values[MSID][index]
-
-    ok = ~bad
-    records = [dt_times[ok].copy()]
-    for msid in msids:
-        records.append(values[msid][ok].copy())
-
-    out = np.rec.fromarrays(records, titles=['time'] + msids)
-    return out
-
-
-def fetch_arrays(start, stop, msids):
-    """
-    Fetch data for ``msids`` from the telemetry archive as arrays.
-
-    This routine is deprecated and is retained only for back-compatibility with
-    old plotting analysis scripts.
-
-    The telemetry values are returned in three dictionaries: ``times``,
-    ``values``, and ``quals``.  Each of these dictionaries contains key-value
-    pairs for each of the input ``msids``.
-
-    :param start: start date of telemetry (Chandra.Time compatible)
-    :param stop: stop date of telemetry
-    :param msids: list of MSIDs (case-insensitive)
-
-    :returns: times, values, quals
-    """
-    times = {}
-    values = {}
-    quals = {}
-
-    for msid in msids:
-        m = MSID(msid, start, stop)
-        times[msid] = m.times
-        values[msid] = m.vals
-        quals[msid] = m.bads
-
-    return times, values, quals
-
-
-def fetch_array(start, stop, msid):
-    """
-    Fetch data for single ``msid`` from the telemetry archive as an array.
-
-    This routine is deprecated and is retained only for back-compatibility with
-    old plotting analysis scripts.
-
-    The telemetry values are returned in three arrays: ``times``, ``values``,
-    and ``quals``.
-
-    :param start: start date of telemetry (Chandra.Time compatible)
-    :param stop: stop date of telemetry
-    :param msid: MSID (case-insensitive)
-
-    :returns: times, values, quals
-    """
-
-    m = MSID(msid, start, stop)
-    times = m.times
-    values = m.vals
-    quals = m.bads
-
-    return times, values, quals
-
-
 class memoized(object):
     """Decorator that caches a function's return value each time it is called.
     If called later with the same arguments, the cached value is returned, and
