+++ setup.py
@@ -12,7 +12,7 @@ if platform.system() in ('Darwin', 'Linux'):
           author='Tom Aldcroft',
           description='Modules supporting Ska engineering telemetry archive',
           author_email='aldcroft@head.cfa.harvard.edu',
-          py_modules=['sken.fetch', 'sken.converters', 'sken.utils'],
+          py_modules=['sken.fetch', 'sken.converters', 'sken.utils', 'sken.fetch_old'],
           version=version,
           zip_safe=False,
           packages=['sken', 'sken.derived', 'sken.tests'],
@@ -24,7 +24,8 @@ setup(name='Ska.engarchive',
       author='Tom Aldcroft',
       description='Modules supporting Ska engineering telemetry archive',
       author_email='aldcroft@head.cfa.harvard.edu',
-      py_modules=['Ska.engarchive.fetch', 'Ska.engarchive.converters', 'Ska.engarchive.utils'],
+      py_modules=['Ska.engarchive.fetch', 'Ska.engarchive.converters', 'Ska.engarchive.utils',
+                  'Ska.engarchive.fetch_old'],
       version=version,
       zip_safe=False,
       namespace_packages=['Ska'],
